requirements.txt
# Backend Framework
fastapi==0.115.9
uvicorn[standard]==0.37.0

# Frontend Framework
streamlit==1.34.0

# LLM and Vector DB
openai==2.3.0          # OpenAI client, also used by Qianfan for API structure
qianfan==0.3.15        # Baidu Qianfan SDK for models and embeddings
chromadb==1.0.3       # Vector database client

# Document Processing
pypdf2==3.0.1          # For reading PDF files
python-docx==1.2.0     # For reading .docx Word files

# HTTP Requests (for frontend to backend communication)
requests==2.32.5

# Data Validation (dependency of FastAPI, good to specify)
pydantic==1.10.15